+++ src/test/java/racingcar/CarTest.java
@@ -4,7 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.engine.model.Car;
+import racingcar.engine.domain.Car;
 
 public class CarTest {
